Added a naive BloomFilter
Added a BloomFilter to amortise the costs of scanning the
ReferenceCounter for each deallocation. This builds a BloomFilter
for all the elements contained in the ReferenceCounter, so we
can check multiple elements with a single scan.

The BloomFilter is very naive as it just uses the four bytes of the
hash as four bit indexs.  It does not adapt its size to the number
of elements, or how its indexs are calculated.